[ICU-6930] IBM i updates: Merging r26624 through r26636  of /icu/branches/srl/i6930 X-Commit-URL: https://ssl.icu-project.org/trac/changeset/26637 Former-commit-id: 496eeea539754d2561db7c5225c8d1d8ad3a9192

+++ icu4c/source/i18n/decimfmt.cpp
@@ -2710,7 +2710,8 @@ DecimalFormat::ERoundingMode DecimalFormat::getRoundingMode() const {
 void DecimalFormat::setRoundingMode(ERoundingMode roundingMode) {
     fRoundingMode = roundingMode;
     if (fRoundingIncrement == NULL) {
-        setRoundingIncrement(pow(10.0, -getMaximumFractionDigits()));
+        /* cast to double to avoid ambiguous pow() overloads */
+        setRoundingIncrement(pow(10.0, (double)-getMaximumFractionDigits()));
     }
 }
